Fix to ensure that we always compute nice log()
It seems that on the instances we are running, the values are getting
very very very low. This can result in log() getting confused and
thinking we are asking something impossible, which makes the whole
algorithm fail miserably. Quick hacking shows that values below 1e-323
are considered as too low to be usables for log(), so we cap at this
level.